+++ components/SeriesPosts.php
@@ -2,7 +2,7 @@
 
 namespace GinoPane\BlogTaxonomy\Components;
 
-use RainLab\Blog\Models\Post;
+use Winter\Blog\Models\Post;
 use GinoPane\BlogTaxonomy\Plugin;
 use GinoPane\BlogTaxonomy\Models\Series;
 use GinoPane\BlogTaxonomy\Classes\PostListAbstract;