+++ models/Tag.php
@@ -2,14 +2,14 @@
 
 namespace GinoPane\BlogTaxonomy\Models;
 
-use Str;
 use System\Models\File;
-use RainLab\Blog\Models\Post;
+use Winter\Blog\Models\Post;
+use Winter\Storm\Support\Str;
 use GinoPane\BlogTaxonomy\Plugin;
-use October\Rain\Database\Builder;
-use October\Rain\Database\Traits\Sluggable;
-use October\Rain\Database\Traits\Validation;
-use October\Rain\Database\Relations\MorphToMany;
+use Winter\Storm\Database\Builder;
+use Winter\Storm\Database\Traits\Sluggable;
+use Winter\Storm\Database\Traits\Validation;
+use Winter\Storm\Database\Relations\MorphToMany;
 
 /**
  * Class Tag
@@ -44,7 +44,7 @@ class Tag extends ModelAbstract
      *
      * @var array
      */
-    public $implement = ['@RainLab.Translate.Behaviors.TranslatableModel'];
+    public $implement = ['@Winter.Translate.Behaviors.TranslatableModel'];
 
     /**
      * Translatable properties, indexed property will be available in queries